requirements.txt
# Core ML dependencies
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.26.0
transformers>=4.37.0

# Image generation dependencies
diffusers>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0

# Model format handlers
safetensors>=0.4.0
huggingface-hub>=0.19.0
ctransformers>=0.2.0

# Optional optimizations
xformers>=0.0.23
bitsandbytes>=0.41.1
